MEC-1249 : prepare-commit-msg hook to prepend JIRA to commit title
Do you remember which ticket you're currently working on? Probably not,
but fear not for this prepare-commit-msg hook is here to help!

This hook (https://git-scm.com/docs/githooks#_prepare_commit_msg) is
invoked after preparing the default log message (i.e. loading a
potential commit template or '-m' value) and before opening the editor. It allows us
to manipulate the entire commit message (i.e. its title and
description).

The JIRA ticket is extracted from the current branch and prepended to
the first line of the commit message. We also add ' : ' since this is
required in PR titles of certain repos and GitHub will use the title of
the first commit of a branch as the PR title.

"$1" represents the file in which the commit message as been placed,
either from reading from the git commit template or from the value of
the '-m' option.

Tests:
- 'git commit -m "<MSG>"' - JIRA ticket prepended
- 'git commit + editor' - JIRA ticket prepended
- and you guest it this commit was done using this hook!

Testing locally:
- make a simple modification
- commit the changed file
- remove the test commit
e.g. `echo $USER >> .gitignore && git commit -m "it's just a test Bert" .gitignore && git reset --hard HEAD^`